+++ integration_tests/kafka/conftest.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 
 import pytest
 
-from .utils import KafkaTestContext
+from .utils import KafkaTestContext, MqttTestContext
 
 
 @pytest.fixture
@@ -12,3 +12,9 @@ def kafka_context():
     kafka_context = KafkaTestContext()
     yield kafka_context
     kafka_context.teardown()
+
+
+@pytest.fixture
+def mqtt_context():
+    mqtt_context = MqttTestContext()
+    yield mqtt_context